# docs/swagger.yaml
definitions:
  dto.CreateOneCategoryRequest:
    properties:
      description:
        maxLength: 225
        minLength: 0
        type: string
      name:
        maxLength: 255
        minLength: 1
        type: string
      parent_id:
        type: integer
    required:
      - name
    type: object
  dto.CreateOnePostRequest:
    properties:
      category_ids:
        items:
          type: integer
        type: array
      contentMarkdown:
        type: string
      image:
        type: string
      title:
        maxLength: 225
        minLength: 1
        type: string
      visibility:
        default: private
        type: string
    required:
      - contentMarkdown
      - title
    type: object
  dto.DeleteOnePostRequest:
    properties:
      id:
        type: integer
    required:
      - id
    type: object
...
  dto.GetAllCategoriesResponse:
    properties:
      categories:
        items:
          $ref: "#/definitions/dto.GetOneCategoryResponse"
        type: array
    type: object
  dto.GetAllPostsResponse:
    description: 获取全部文章响应
    properties:
      category_ids:
        items:
          type: integer
        type: array
      contentHtml:
        type: string
      contentMarkdown:
        type: string
      id:
        type: integer
      image:
        type: string
      title:
        type: string
      visibility:
        type: string
    type: object
  dto.GetOneCategoryResponse:
    properties:
      children:
        items:
          $ref: "#/definitions/dto.GetOneCategoryResponse"
        type: array
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
    type: object
...
  dto.UpdateOneCategoryRequest:
    properties:
      children:
        items:
          $ref: "#/definitions/dto.GetOneCategoryResponse"
        type: array
      description:
        maxLength: 255
        minLength: 0
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        maxLength: 255
        minLength: 1
        type: string
      parent_id:
        type: integer
      path:
        type: string
    required:
      - description
      - id
      - name
    type: object
  dto.UpdateOnePostRequest:
    properties:
      category_ids:
        items:
          type: integer
        type: array
      contentMarkdown:
        type: string
      id:
        type: integer
      image:
        type: string
      title:
        maxLength: 255
        minLength: 0
        type: string
      visibility:
        default: private
        type: string
    required:
      - category_ids
      - id
    type: object
  vo.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      requestId: {}
      timestamp: {}
    type: object
info:
  contact: {}
paths:
...
  /category:
    get:
      consumes:
        - application/json
      description: 获取类目树
      produces:
        - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    items:
                      $ref: "#/definitions/dto.GetAllCategoriesResponse"
                    type: array
                type: object
        "500":
          description: 服务器错误
          schema:
            $ref: "#/definitions/vo.Result"
      summary: 获取类目树
      tags:
        - 类目
    post:
      consumes:
        - application/json
      description: 创建新的类目
      parameters:
        - description: 创建类目请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.CreateOneCategoryRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    $ref: "#/definitions/dto.GetOneCategoryResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/vo.Result"
      security:
        - BearerAuth: []
      summary: 创建类目
      tags:
        - 类目
  /category/{id}:
    delete:
      consumes:
        - application/json
      description: 根据类目 ID 删除指定类目
      parameters:
        - description: 类目ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    $ref: "#/definitions/dto.GetOneCategoryResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/vo.Result"
        "404":
          description: 类目不存在
          schema:
            $ref: "#/definitions/vo.Result"
        "500":
          description: 服务器错误
          schema:
            $ref: "#/definitions/vo.Result"
      security:
        - BearerAuth: []
      summary: 删除类目
      tags:
        - 类目
    get:
      consumes:
        - application/json
      description: 根据类目 ID 获取单个类目的详细信息
      parameters:
        - description: 类目ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    $ref: "#/definitions/dto.GetOneCategoryResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/vo.Result"
        "404":
          description: 类目不存在
          schema:
            $ref: "#/definitions/vo.Result"
      summary: 获取单个类目详情
      tags:
        - 类目
    put:
      consumes:
        - application/json
      description: 更新已存在的类目信息
      parameters:
        - description: 类目ID
          in: path
          name: id
          required: true
          type: integer
        - description: 更新类目请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.UpdateOneCategoryRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    $ref: "#/definitions/dto.GetOneCategoryResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/vo.Result"
        "404":
          description: 类目不存在
          schema:
            $ref: "#/definitions/vo.Result"
        "500":
          description: 服务器错误
          schema:
            $ref: "#/definitions/vo.Result"
      security:
        - BearerAuth: []
      summary: 更新类目
      tags:
        - 类目
  /category/{id}/children:
    get:
      consumes:
        - application/json
      description: 根据类目 ID 获取子类目树
      parameters:
        - description: 类目ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    items:
                      $ref: "#/definitions/dto.GetOneCategoryResponse"
                    type: array
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/vo.Result"
        "404":
          description: 类目不存在
          schema:
            $ref: "#/definitions/vo.Result"
        "500":
          description: 服务器错误
          schema:
            $ref: "#/definitions/vo.Result"
      summary: 获取子类目树
      tags:
        - 类目
  /post/createOnePost:
    post:
      consumes:
        - application/json
      description: 创建新的文章，支持 Markdown 格式内容，系统会自动转换为 HTML
      parameters:
        - description: 创建文章请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.CreateOnePostRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    $ref: "#/definitions/dto.GetAllPostsResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/vo.Result"
        "500":
          description: 服务器错误
          schema:
            $ref: "#/definitions/vo.Result"
      security:
        - BearerAuth: []
      summary: 创建文章
      tags:
        - 文章
  /post/deleteOnePost:
    post:
      consumes:
        - application/json
      description: 根据文章 ID 删除指定文章
      parameters:
        - description: 删除文章请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.DeleteOnePostRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: "#/definitions/vo.Result"
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/vo.Result"
        "404":
          description: 文章不存在
          schema:
            $ref: "#/definitions/vo.Result"
        "500":
          description: 服务器错误
          schema:
            $ref: "#/definitions/vo.Result"
      security:
        - BearerAuth: []
      summary: 删除文章
      tags:
        - 文章
  /post/getAllPost:
    get:
      consumes:
        - application/json
      description: 获取所有的文章列表，按创建时间倒序排序
      produces:
        - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    items:
                      $ref: "#/definitions/dto.GetAllPostsResponse"
                    type: array
                type: object
        "500":
          description: 服务器错误
          schema:
            $ref: "#/definitions/vo.Result"
      summary: 获取文章列表
      tags:
        - 文章
  /post/updateOnePost:
    post:
      consumes:
        - application/json
      description: 更新已存在的文章内容
      parameters:
        - description: 更新文章请求参数
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/dto.UpdateOnePostRequest"
      produces:
        - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
              - $ref: "#/definitions/vo.Result"
              - properties:
                  data:
                    $ref: "#/definitions/dto.GetAllPostsResponse"
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: "#/definitions/vo.Result"
        "404":
          description: 文章不存在
          schema:
            $ref: "#/definitions/vo.Result"
        "500":
          description: 服务器错误
          schema:
            $ref: "#/definitions/vo.Result"
      security:
        - BearerAuth: []
      summary: 更新文章
      tags:
        - 文章
...
swagger: "2.0"