# pkg/serve/controller/category/dto/get_all_categories_response.go
package dto

// GetAllCategoriesResponse 获取所有类目响应
type GetAllCategoriesResponse struct {
	Categories []*GetOneCategoryResponse `json:"categories" xml:"categories" form:"categories" query:"categories"`
}